Adding pause/resume support to the async loader

+++ src/resource/AsyncLoader.cpp
@@ -20,28 +20,24 @@ AsyncLoader::~AsyncLoader()
 {
 	stop();
 
-	if(m_head != nullptr)
+	if(!m_taskQueue.isEmpty())
 	{
-		ANKI_ASSERT(m_tail != nullptr);
 		ANKI_LOGW("Stoping loading thread while there is work to do");
 
-		AsyncLoaderTask* task = m_head;
-
-		do
+		while(!m_taskQueue.isEmpty())
 		{
-			AsyncLoaderTask* next = task->m_next;
+			AsyncLoaderTask* task = &m_taskQueue.getFront();
+			m_taskQueue.popFront();
 			m_alloc.deleteInstance(task);
-			task = next;
-		} while(task != nullptr);
+		}
 	}
 }
 
 //==============================================================================
-Error AsyncLoader::create(const HeapAllocator<U8>& alloc)
+void AsyncLoader::init(const HeapAllocator<U8>& alloc)
 {
 	m_alloc = alloc;
 	m_thread.start(this, threadCallback);
-	return ErrorCode::NONE;
 }
 
 //==============================================================================
@@ -50,12 +46,32 @@ void AsyncLoader::stop()
 	{
 		LockGuard<Mutex> lock(m_mtx);
 		m_quit = true;
-
 		m_condVar.notifyOne();
 	}
 
 	Error err = m_thread.join();
 	(void)err;
+}
+
+//==============================================================================
+void AsyncLoader::pause()
+{
+	{
+		LockGuard<Mutex> lock(m_mtx);
+		m_paused = true;
+		m_sync = true;
+		m_condVar.notifyOne();
+	}
+
+	m_barrier.wait();
+}
+
+//==============================================================================
+void AsyncLoader::resume()
+{
+	LockGuard<Mutex> lock(m_mtx);
+	m_paused = false;
+	m_condVar.notifyOne();
 }
 
 //==============================================================================
@@ -72,12 +88,14 @@ Error AsyncLoader::threadWorker()
 
 	while(!err)
 	{
-		AsyncLoaderTask* task;
+		AsyncLoaderTask* task = nullptr;
+		Bool quit = false;
+		Bool sync = false;
 
 		{
 			// Wait for something
 			LockGuard<Mutex> lock(m_mtx);
-			while(m_head == nullptr && m_quit == false)
+			while((m_taskQueue.isEmpty() || m_paused) && !m_quit && !m_sync)
 			{
 				m_condVar.wait(m_mtx);
 			}
@@ -85,32 +103,57 @@ Error AsyncLoader::threadWorker()
 			// Do some work
 			if(m_quit)
 			{
-				break;
+				quit = true;
 			}
-
-			task = m_head;
-
-			// Update the queue
-			if(m_head->m_next == nullptr)
+			else if(m_sync)
 			{
-				ANKI_ASSERT(m_tail == m_head);
-				m_head = m_tail = nullptr;
+				sync = true;
+				m_sync = false;
 			}
 			else
 			{
-				m_head = m_head->m_next;
+				task = &m_taskQueue.getFront();
+				m_taskQueue.popFront();
 			}
 		}
 
-		// Exec the task
-		err = (*task)();
-		if(err)
+		if(quit)
 		{
-			ANKI_LOGE("Async loader task failed");
+			break;
 		}
+		else if(sync)
+		{
+			m_barrier.wait();
+		}
+		else
+		{
+			// Exec the task
+			ANKI_ASSERT(task);
+			AsyncLoaderTaskContext ctx;
+			err = (*task)(ctx);
+			if(err)
+			{
+				ANKI_LOGE("Async loader task failed");
+			}
 
-		// Delete the task
-		m_alloc.deleteInstance(task);
+			// Do other stuff
+			if(ctx.m_resubmitTask)
+			{
+				LockGuard<Mutex> lock(m_mtx);
+				m_taskQueue.pushBack(task);
+			}
+			else
+			{
+				// Delete the task
+				m_alloc.deleteInstance(task);
+			}
+
+			if(ctx.m_pause)
+			{
+				LockGuard<Mutex> lock(m_mtx);
+				m_paused = true;
+			}
+		}
 	}
 
 	return err;
